requirements.txt
requests~=2.25.1
pandas~=1.2.3
numpy~=1.20.1
Django~=3.1.7
matplotlib~=3.4.0
selenium~=3.141.0
aiohttp~=3.7.4.post0
tqdm~=4.60.0
djangorestframework~=3.12.4